move WorkCombinedAudioDerivatives presenter to OralHistory::Admin::CombinedAudioDerivativesComponent
