--- app/views/admin/works/_combined_audio_derivatives.html.erb
@@ -1,75 +1,0 @@
-<%# view is WorkCombinedAudioDerivatives #%>
-<div class="card bg-light mb-3">
-  <h2 class="card-header h3">Combined Audio Derivative Files</h2>
-  <div class="card-body">
-
-
-
-    <% if view.derivatives_up_to_date? %>
-        <p class="alert-success p-3">
-          <i class="fa fa fa-thumbs-up pr-1" aria-hidden="true"></i>
-            The combined audio derivatives, created from <%= view.work_available_members_count %> published audio segments, are up to date. Their fingerprint is <code class="border"><%= view.combined_audio_fingerprint%></code>.
-        </p>
-        <ul>
-          <% if view.combined_mp3_audio.present? %>
-            <li>
-              Use the following mp3 link as the OHMS Media URL:<br/>
-              <small>
-                <%= link_to view.combined_mp3_audio,  view.combined_mp3_audio %>
-              </small>
-            </li>
-          <% end %>
-          <% if view.combined_webm_audio.present? %>
-            <li>
-              The corresponding webm link is provided for reference only:<br/>
-              <small>
-                <%= link_to view.combined_webm_audio, view.combined_webm_audio %>
-              </small>
-            </li>
-          <% end %>
-        </ul>
-    <% elsif ! view.work_available_members? %>
-      <p class="alert-info p-3">
-        This oral history doesn't have any published audio segments associated with it, so has no combined audio derivatives.
-      </p>
-    <% else %>
-      <div class="alert-danger p-3 mb-2">
-        <p>
-          <i class="fa fa fa-thumbs-down pr-1" aria-hidden="true"></i>
-          This oral history does not have up to date combined audio derivatives for the <%= view.work_available_members_count %> published audio segment(s).
-          No combined audio player can be shown!
-        </p>
-        <% if view.combined_audio_fingerprint %>
-          <ul>
-            <li>Actual fingerprint: <%= view.combined_audio_fingerprint %></li>
-            <li>Required fingerprint: <%= view.current_required_fingerprint %></li>
-          </ul>
-        <% else %>
-          <p>No combined derivative exists.</p>
-        <% end %>
-      </div>
-    <% end %>
-
-    <% if view.work_available_members? %>
-      <p>
-        <%= link_to "Generate combined audio derivatives", create_combined_audio_derivatives_admin_work_path(model), method: "put", class: "btn btn-primary" %>
-      </p>
-    <% end %>
-
-    <% if view.show_in_progress_status? %>
-      <p class="<%= model&.oral_history_content&.failed? ? 'alert-danger' : 'alert-info' %>  p-3">
-        Attempting to create combined audio derivatives.
-        Job status:
-        <code><%= model&.oral_history_content&.combined_audio_derivatives_job_status %></code>
-        as of
-        <%= l model&.oral_history_content&.combined_audio_derivatives_job_status_changed_at, format: :admin %>
-      </p>
-    <% end %>
-
-    <% if model&.oral_history_content&.succeeded? && view.job_status_recently_changed? %>
-      <p class="alert-info p-3">
-        Combined audio derivatives were recently created or recreated. The job completed successfully at <%= l model&.oral_history_content&.combined_audio_derivatives_job_status_changed_at, format: :admin %>.
-      </p>
-    <% end %>
-  </div>
-</div>
