+++ app/views/admin/works/show.html.erb
@@ -142,7 +142,7 @@
 
   <% if @work.is_oral_history? %>
     <div class="tab-pane" id="nav-oral-histories" role="tabpanel" aria-labelledby="nav-oral-histories-tab">
-      <%= WorkCombinedAudioDerivatives.new(@work).display %>
+      <%= render OralHistory::Admin::CombinedAudioDerivativesComponent.new(work: @work) %>
       <%= render "oral_history_biography",  work: @work %>
       <%= render "ohms_xml",                work: @work %>
       <%= render "oral_history_transcript", work: @work %>
